Use static_cast

+++ src/moodbar/moodbarbuilder.cpp
@@ -176,7 +176,7 @@ QByteArray MoodbarBuilder::Finish(int width) {
   for (int i = 0; i < width; ++i) {
     Rgb rgb;
     const int start = i * frames_.count() / width;
-    const int end = std::max((i + 1) * frames_.count() / width, start + 1);
+    const int end = std::max((i + 1) * static_cast<int>(frames_.count()) / width, start + 1);
 
     for (int j = start; j < end; j++) {
       const Rgb& frame = frames_[j];
